error catching and information for user

+++ frontend/src/app/services/auth/auth.service.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
 import { MessageInfo } from '../../models/MessageInfo';
-import { HttpClient, HttpErrorResponse, HttpHeaders } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { API_URL } from '../../config';
-import { catchError } from 'rxjs/operators';
 import { Router } from '@angular/router';
 import { Authorities } from '../../models/database/entites/Authorities';
 
@@ -27,36 +26,45 @@ export class AuthService {
 	}
 
 	async getOne(url: string) {
-		return await this.http.get<MessageInfo>(this.URL + url, this.setHeaders())
-		.pipe(
-			catchError(this.handleError)
-		).toPromise();
+	  try {
+      return await this.http.get<MessageInfo>(this.URL + url, this.setHeaders()).toPromise();
+    } catch (e) {
+      sessionStorage.clear();
+      sessionStorage.setItem('ROLE_ADMIN', 'false');
+      sessionStorage.setItem('ROLE_EMPLOYEE', 'false');
+      sessionStorage.setItem('authenticated', 'false');
+      sessionStorage.setItem('isSignedCorrectly', 'false');
+      return null;
+    }
 	}
 
 	async loginUser() {
-		const response: MessageInfo = await this.getOne('user') as MessageInfo;
-		if (response.object) {
-			sessionStorage.setItem('authenticated', 'true');
-			const roles: Authorities[] = response.object.authorities;
-			for (let i = 0; i < roles.length; i++) {
-				if (roles[i].authority === 'ROLE_ADMIN') {
-					sessionStorage.setItem('ROLE_ADMIN', 'true');
-				} else if (roles[i].authority === 'ROLE_EMPLOYEE') {
-					sessionStorage.setItem('ROLE_EMPLOYEE', 'true');
-				}
-			}
 
-			this.router.navigateByUrl('/homepage')
-			.then(() => {
-				location.reload();
-			});
+    const response: MessageInfo = await this.getOne('user') as MessageInfo;
+    if (response && response.object) {
+      sessionStorage.setItem('authenticated', 'true');
+      const roles: Authorities[] = response.object.authorities;
+      for (let i = 0; i < roles.length; i++) {
+        if (roles[i].authority === 'ROLE_ADMIN') {
+          sessionStorage.setItem('ROLE_ADMIN', 'true');
+        } else if (roles[i].authority === 'ROLE_EMPLOYEE') {
+          sessionStorage.setItem('ROLE_EMPLOYEE', 'true');
+        }
+      }
 
-		} else {
-			sessionStorage.clear();
-			sessionStorage.setItem('ROLE_ADMIN', 'false');
-			sessionStorage.setItem('ROLE_EMPLOYEE', 'false');
-			sessionStorage.setItem('authenticated', 'false');
-		}
+      this.router.navigateByUrl('/homepage')
+        .then(() => {
+          sessionStorage.setItem('isSignedCorrectly', 'true');
+          location.reload();
+        });
+
+    } else {
+      sessionStorage.clear();
+      sessionStorage.setItem('ROLE_ADMIN', 'false');
+      sessionStorage.setItem('ROLE_EMPLOYEE', 'false');
+      sessionStorage.setItem('authenticated', 'false');
+      sessionStorage.setItem('isSignedCorrectly', 'false');
+    }
 	}
 
 	logoutUser() {
@@ -64,19 +72,11 @@ export class AuthService {
 		sessionStorage.setItem('ROLE_ADMIN', 'false');
 		sessionStorage.setItem('ROLE_EMPLOYEE', 'false');
 		sessionStorage.setItem('authenticated', 'false');
+    sessionStorage.setItem('isSignedCorrectly', 'true');
 
 		this.router.navigateByUrl('/homepage')
 		.then(() => {
 			location.reload();
 		});
-	}
-
-
-	private handleError(error: HttpErrorResponse) {
-		sessionStorage.clear();
-		sessionStorage.setItem('ROLE_ADMIN', 'false');
-		sessionStorage.setItem('ROLE_EMPLOYEE', 'false');
-		sessionStorage.setItem('authenticated', 'false');
-		return null;
 	}
 }
